events manager refactoring
